backend/requirements.txt
# 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# 데이터베이스
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1

# OpenSearch
opensearch-py==2.4.2
opensearch-dsl==2.1.0

# LLM & RAG
langchain==0.1.0
langgraph==0.0.20
openai==1.3.7
sentence-transformers==2.2.2
tiktoken==0.5.2

# 작업 큐
celery==5.3.4
celery[redis]==5.3.4

# 유틸리티
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
emails==0.6
httpx==0.25.2

# 모니터링 & 로깅
prometheus-client==0.19.0
python-json-logger==2.0.7

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# 데이터 처리
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2

# API 문서화
mkdocs==1.5.3
mkdocs-material==9.5.2